# lib/api-zod/src/generated/types/accountUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AccountUser {
  id: string;
  name: string;
  email: string;
}