# lib/api-zod/src/generated/types/authState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AccountUser } from './accountUser';
import type { SemesterSimulationInput } from './semesterSimulationInput';

export interface AuthState {
  authenticated: boolean;
  user: AccountUser | null;
  profile: SemesterSimulationInput | null;
}