# lib/api-zod/src/generated/types/opportunityMatchInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OpportunityMatchInput {
  /** @minimum 0 */
  tuitionSupportAmount: number;
  /** @minimum 0 */
  monthlyIncomeAmount: number;
  /** @minimum 0 */
  semesterIncomeAmount: number;
  /** @minimum 0 */
  currentGap: number;
  /**
     * @minimum 1
     * @maximum 12
     */
  semesterDuration: number;
  /** @minLength 1 */
  major: string;
}