# lib/api-zod/src/generated/types/opportunityMatchResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Opportunity } from './opportunity';
import type { OpportunityTargets } from './opportunityTargets';

export interface OpportunityMatchResult {
  targets: OpportunityTargets;
  opportunities: Opportunity[];
  disclaimer: string;
}