# lib/api-zod/src/generated/types/optimizerCandidate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OpportunityTargets } from './opportunityTargets';
import type { OptimizerChanges } from './optimizerChanges';
import type { SemesterSimulationInput } from './semesterSimulationInput';
import type { SemesterSimulationResult } from './semesterSimulationResult';

export interface OptimizerCandidate {
  profile: SemesterSimulationInput;
  result: SemesterSimulationResult;
  changes: OptimizerChanges;
  targets: OpportunityTargets;
  stabilized: boolean;
  /** @minimum 0 */
  gap: number;
}