# lib/api-zod/src/generated/types/optimizerIntervention.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OptimizerChanges } from './optimizerChanges';
import type { SemesterSimulationResult } from './semesterSimulationResult';

export interface OptimizerIntervention {
  kind: string;
  label: string;
  changes: OptimizerChanges;
  isolatedResult: SemesterSimulationResult;
  gapReduction: number;
  scoreChange: number;
  stabilized: boolean;
  whyThisHelps: string;
}