# lib/api-zod/src/generated/types/semesterSimulationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TuitionPayment } from './tuitionPayment';

export interface SemesterSimulationInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  major: string;
  /** @minimum 1 */
  creditHours: number;
  /** @minimum 0 */
  costPerCredit: number;
  /**
     * @minimum 1
     * @maximum 12
     */
  semesterDuration: number;
  currentBalance: number;
  /** @minimum 0 */
  monthlyIncome: number;
  /** @minimum 0 */
  transportation: number;
  /** @minimum 0 */
  food: number;
  /** @minimum 0 */
  otherExpenses: number;
  tuitionPayments: TuitionPayment[];
}